UPSTREAM: <carry>: openshift: Stop asigning random az
This changes default behaviour to be deterministic and to give explicit control to the user over using zones or not and which one to choose instead of deploying to a random a zone by default. We want to provide as similar as possible user experience between cloud providers. Although in the future actuators might get smarter or/and integrate with provider specific sets/asg we design and assume for them being dumb simple and control is given to the user via inputs.