UPSTREAM: <carry>: openshift: Update machine status and spec outside of a machine scope
With machine scope we have no control of when a machine status or spec get updated.
Also, we need to update a status/spec multiple times before end of an operation (e.g. Create)
is reached. Machine scope also contains a lot of boilerplate (e.g. cluster bits) which
we don't need. The next step is to remove machine scope completely.