UPSTREAM: <carry>: openshift: do not randomize public ip name, error instead
Generated ip public name has to be the same all the time since the name is generated
in two different operations (Create and Delete) which are independent of each other.
Unless the public ip name is stored in machine status, the delete has to get
the same name as the create operation to delete the resource from Azure.