UPSTREAM: <carry>: openshift: Assert VM exists if VM state is Creating
Currently, when a VM is being created in Azure, we first check whether the VM exists and then attempt to create it if not.

On the first attempt to create, we always get an asynchronous time out:
```
E0709 16:44:25.752572       1 actuator.go:78] Machine error: failed to reconcile machine "jspeed-test-8cnpg-worker-centralus3-g6htz"s: failed to create vm jspeed-test-8cnpg-worker-centralus3-g6htz: failed to create or get machine: compute.VirtualMachinesCreateOrUpdateFuture: asynchronous operation has not completed
```

This will cause the Machine to be requeued. Currently, because Exists does not determine the machine to exist, we attempt to create again and get the following error:

```
E0709 16:44:26.642473       1 actuator.go:78] Machine error: failed to reconcile machine "jspeed-test-8cnpg-worker-centralus3-g6htz"s: failed to create vm jspeed-test-8cnpg-worker-centralus3-g6htz: vm jspeed-test-8cnpg-worker-centralus3-g6htz is still in provisioning state Creating, reconcile
```

The VM exists immediately after the first call to create the VM, but we are attempting to recreate it becuase `Exists` claims that it does not exist. This can lead to issues if there is a transient error after the first `Create` but before the VM becomes `Running`. We could see an error, determine the creation failed and move the Machine to the `Failed` phase.

If this happens, the VM will still start, but because the Machine is Failed, we do not track it, we do not remove it if the Machine is deleted. Therefore we can leak VMs.

This PR fixes `Exists` such that if the VM exists on the API, but is in the `Creating` phase, it is considered to exist. Now we do not attempt to create the VM more than once and, if the first VM creation is successful (even with the async error), the Machine is considered to exist, so we will not fail a Machine while the VM is in the `Creating` phase.