UPSTREAM: <carry>: openshift: vm: Add AdditionalUnattendContent on Windows machines
This commit adds functionality for automatically running
C:\AzureData\CustomData.bin on first boot.
This file is provided via the user data secret,
and contains init scripts.

The Windows Machine Config Operator relies on the user data script being
executed automatically on the Windows VM to set up SSH access.

On AWS, it is invoked automatically after provisioning.
This commit is intended to align Azure with that behavior.